api/pyproject.toml
[project]
name = "aerospike-cluster-manager-api"
version = "0.1.0"
description = "Aerospike Cluster Manager Backend API"
readme = "README.md"
authors = [
    { name = "aerospike-ce-ecosystem", email = "aerospike-ce-ecosystem@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aerospike-py[otel]>=0.6.4",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.0",
    "kubernetes>=31.0.0",
    "mcp>=1.0,<2",
    "python-json-logger>=3.0.0",
    "pyyaml>=6.0",
    "slowapi>=0.1.9",
    "sse-starlette>=2.0.0",
    "swagger-ui-bundle>=1.1.0",
    "uvicorn[standard]>=0.34.0",
    # OpenTelemetry — disabled at runtime via OTEL_SDK_DISABLED=true (default).
    # All exporter/sampler/resource configuration goes through OTel SDK
    # standard environment variables; no extra wrappers in this codebase.
    # No upper bound on minor versions — opentelemetry releases SDK and
    # instrumentation packages on the same cadence and renovate can bump
    # them in lockstep.
    "opentelemetry-api>=1.30.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-otlp>=1.30.0",
    "opentelemetry-instrumentation-fastapi>=0.51b0",
    "opentelemetry-instrumentation-asyncpg>=0.51b0",
    "opentelemetry-instrumentation-logging>=0.51b0",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.3,<4",
]

[tool.uv]
# Ensure uv locks wheels for CI (Linux x86_64) and container (Linux aarch64) platforms,
# not just the developer's macOS machine
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'linux' and platform_machine == 'aarch64'",
    "sys_platform == 'darwin'",
]

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["aerospike_cluster_manager_api"]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
    "pyright>=1.1.408",
    "pytest>=9.0.3",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
    "pytest-cov>=5.0.0",
    "testcontainers[postgres]>=4.0.0",
]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
pythonPlatform = "All"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
